api_project/requirements.txt

asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
django-restframework==0.0.1
djangorestframework==3.15.1
psycopg2==2.9.9
sqlparse==0.4.4
typing-extensions==4.11.0
tzdata==2024.1